Revert "Making scripting engine plugable"
