+++ components/mediators/datamapper/org.wso2.carbon.mediator.datamapper.engine/src/main/java/org/wso2/datamapper/engine/core/MappingHandler.java
@@ -15,15 +15,13 @@
  */
 package org.wso2.datamapper.engine.core;
 
-import org.apache.avro.generic.GenericRecord;
-import org.json.JSONException;
-import org.wso2.datamapper.engine.core.exceptions.JSException;
-import org.wso2.datamapper.engine.core.executors.ScriptExecutorFactory;
-import org.wso2.datamapper.engine.core.executors.ScriptExecutorType;
-import org.wso2.datamapper.engine.inputAdapters.InputDataReaderAdapter;
-
 import java.io.IOException;
 import java.io.InputStream;
+
+import org.apache.avro.generic.GenericRecord;
+import org.json.JSONException;
+import org.wso2.datamapper.engine.core.MappingResourceLoader.JSException;
+import org.wso2.datamapper.engine.inputAdapters.InputDataReaderAdapter;
 
 
 public class MappingHandler {
@@ -35,8 +33,7 @@ public class MappingHandler {
 		//InputDataReaderAdapter inputReader = new CsvInputReader();
 		inputReader.setInputMsg(inputMsg);
 	    GenericRecord inputRecord = inputReader.getInputRecord(resourceModel.getInputSchema());
-		IScriptExecutor scriptExecutor = ScriptExecutorFactory.getScriptExecutor(ScriptExecutorType.RHINO);
-		GenericRecord outputRecord = scriptExecutor.executeMapping(resourceModel, inputRecord);
+		GenericRecord outputRecord = FunctionExecuter.execute(resourceModel, inputRecord);
 	    
 		return outputRecord;
