--- components/mediators/datamapper/org.wso2.carbon.mediator.datamapper.engine/src/main/java/org/wso2/datamapper/engine/core/exceptions/JSException.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2016 WSO2, Inc. http://www.wso2.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.datamapper.engine.core.exceptions;
-
-/**
- * This exception is thrown when script executors can not execute the script
- */
-public class JSException extends Exception {
-
-
-    private static final long serialVersionUID = 1L;
-
-    private String message = null;
-
-    public JSException() {
-        super();
-    }
-
-    public JSException(String message) {
-        super(message);
-        this.message = message;
-    }
-
-    public JSException(Throwable cause) {
-        super(cause);
-    }
-
-    @Override
-    public String toString() {
-        return message;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
