--- components/mediators/datamapper/org.wso2.carbon.mediator.datamapper.engine/src/main/java/org/wso2/datamapper/engine/core/executors/ScriptExecutorType.java
@@ -1,23 +1,0 @@
-/*
- * Copyright 2016 WSO2, Inc. http://www.wso2.org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wso2.datamapper.engine.core.executors;
-
-/**
- * This enum type represent the types of executors support in data mapper
- */
-public enum ScriptExecutorType {
-    RHINO,NASHORN,J2V8,V8
-}
